backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Background Jobs
redis==5.0.1
rq==1.16.0
celery==5.3.6
apscheduler==3.10.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0

# OAuth
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
msal==1.26.0

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# LLM API
openai==1.10.0
google-generativeai==0.3.2

# Email Parsing & Processing
beautifulsoup4==4.12.3
lxml==5.1.0
html2text==2024.2.26
email-validator==2.1.0

# Integrations
slack-sdk==3.26.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Storage (AWS S3 / CloudFlare R2)
boto3==1.34.22

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
